==> KEGG/01.KOG_2_uids.ipynb <==
import requests
from bs4 import BeautifulSoup
from pprint import pprint
import regex as re
from multiprocessing import Pool
import sys

base_url = "https://www.kegg.jp/kegg-bin/uniprot_list?ko=" # append KO group to end of url
def transform_int2knumber(knint):
    knstr = str(knint)
    while len(knstr) < 5:
        knstr = "0" + knstr
    return "K" + knstr
not_kog_ids = []
with open("not_knumber.log", "r") as f:
    lines = f.readlines()
    for line in lines:
        knumber = line.strip()
        not_kog_ids.append(knumber)
kog_ids_raw = [transform_int2knumber(k) for k in range(1, 27110)]
kog_ids = [k for k in kog_ids_raw if k not in not_kog_ids]

def extract_uid_htmlrow(rowtext):
    uid = re.findall(r'">(.*)<\/a', rowtext)[0]
    return uid
def extract_descriptor_htmlrow(rowtext, altformat = False):
    if altformat:
        descriptor = re.findall(r'[^\n].*$', descriptor_row)[0].split("\xa0")[-1]
        return descriptor
    descriptor = re.findall(r';\s(.*)', rowtext)[0] 
    return descriptor

def get_uids(kogid): # also gets descriptor for KOG
    uids = []
    url = base_url + kogid
    data = requests.get(url)
    html = BeautifulSoup(data.text, 'html.parser')
    try: 
        descriptor_row = html.div.text
    except:
        with open("not_knumber_anymore.log", "a") as outfile:
            print(kogid, file=outfile)
            return
    print("completed request for {}".format(kogid))
    try:
        descriptor = extract_descriptor_htmlrow(descriptor_row)
    except: # row is formated without ";" separating the gene alias and descriptor
        descriptor = extract_descriptor_htmlrow(descriptor_row, True)
    print("got descriptor")

    all_tds = html.find_all("td")
    for i in range(2, len(html.find_all("td")), 4):
        rowtext = str(all_tds[i])
        uid = extract_uid_htmlrow(rowtext)
        uids.append(uid)

    output = "data/kog_uid/{}.tsv".format(kogid)
    with open(output, "a") as outfile:
        for uid in uids:
            print("\t".join([kogid, uid, descriptor]), file=outfile)
#get_uids("K15032")

section, num_sections = sys.argv[1], sys.argv[2]

for kog_id in kog_ids[:8]:
    get_uids(kog_id)

n = 1094
section = 2
sections = 8

for i in range(0, 1094, int(1094/8)):
    print(i)


kog_ids_100 = kog_ids

num_sections = 10
section = 10

len(kog_ids_100) // num_sections

section_indices = [i for i in range(0, len(kog_ids_100), len(kog_ids_100) // num_sections)]
start_range = section_indices[section-1]
if section == num_sections:
    end_range = len(kog_ids_100)
else:
    end_range = section_indices[section]

start_range, end_range

kog_ids_100[len(kog_ids_100)]

len(kog_ids_100)

a[len(a)]
