==> KEGG/04.logjobs.ipynb <==
#from Bio import SeqIO
import subprocess
import regex as re

section = 3

def check_jobstatus(jobid):
    command = f"curl -i 'https://rest.uniprot.org/idmapping/status/{jobid}'"
    response = subprocess.run(
        command,
        text=True,
        stdout=subprocess.PIPE,
        check=True,
        shell = True)
    return response.stdout

jobid_file = f"/data/luojaa/kegg/batch_jobids/section{section}.tsv"
job_log = f"/data/luojaa/kegg/section{section}.log"
fcount, ncount, rcount = 0,0,0
with open(jobid_file, "r") as f, open(job_log, "a") as logfile:
    lines = f.readlines()
    for line in lines:
        batch, jobid = line.strip().split("\t")
        status_out = check_jobstatus(jobid)
        status = re.findall(r"jobStatus\":\"(.*)\"}", status_out)[0]
        print(status)
        if status =="FINISHED":
            fcount += 1
        elif status == "NEW":
            ncount +=1
        elif status == "RUNNING":
            rcount += 1
    print(",".join([count, ncount, rcount], file=logfile)

def download_fasta(jobid):
    url = f"https://rest.uniprot.org/idmapping/uniref/results/stream/{jobid}?compressed=true&format=fasta"
    destination = f"data/genes/test.{jobid}.fasta"
    command = ["curl", url, "--output", destination]
    response = subprocess.run(
        command,
        text=True,
        stdout=subprocess.PIPE,
        check=True)
