==> KEGG/07.stats_kogs.ipynb <==
import os
from Bio import SeqIO
import pandas as pd
import numpy as np
import ete3

base_stats = "/data/luojaa/uid_stats/"

kogs_df_search = pd.read_csv("/data/luojaa/uid_stats/kogs_df_searchable.tsv", sep = "\t")

kogs_df_func = kogs_df_search[["KOGID", "FUNCTION"]].dropna()
kogs_df_func = kogs_df_func[kogs_df_func["FUNCTION"] != "none"]

kogs_df_func_drep = kogs_df_func.drop_duplicates()

kogs_df_func_drep.to_csv("/data/luojaa/uid_stats/kog_functions.tsv", sep = "\t", index = None)

# # forgot to combine these files earlier.
# base = "/data/luojaa/kegg/kog_uid/"
# kog2uidpaths = [base + tsv for tsv in os.listdir(base)]
# with open("/data/luojaa/uid_stats/uid2descriptors.tsv","w") as outfile:
#     for path in kog2uidpaths:
#         with open(path, "r") as f:
#             for line in f:
#                 kog, uid, descriptor = line.strip().split("\t")
#                 print("\t".join([uid, descriptor]), file=outfile)

#base_stats = "/data/luojaa/uid_stats/"
genes2uid = base_stats + "kegg_genes.mappings.csv"
uid2kogs = base_stats + "uid2kogs.csv"
uid2taxid = base_stats + "uid2taxids.combined.tsv"
uid2descriptor = base_stats + "uid2descriptors.tsv"

genes2uid_df = pd.read_csv(genes2uid).rename(columns={"UNIPROT_ID":"UID"}).set_index("UID")
# missing descriptors for genes w/o UID, and some UIDs lacking descriptors
uid2kogs_df = pd.read_csv(uid2kogs).iloc[:, 1:].rename(columns = {"0":"UID","1":"KOG"}).set_index("UID")
uid2taxid_df = pd.read_csv(uid2taxid, sep = "\t", header = None).rename(columns={0:"UID",1:"TAXID"}).set_index("UID")
uid2desc_df = pd.read_csv(uid2descriptor, sep = "\t", header = None).rename(columns={0:"UID",1:"NAME"}).set_index("UID")

# which KOGs have genes in both euk and prok
uid2kogs = base_stats + "uid2kogs.csv"
uid2kogs_df = pd.read_csv(uid2kogs).iloc[:, 1:].rename(columns = {"0":"UID","1":"KOG"}).set_index("UID")

uid2taxid_iseuk = base_stats + "uid2taxids2iseuk.combined.tsv"
uid2taxid_iseuk_df = pd.read_csv(uid2taxid_iseuk, sep = "\t", header = None).rename(columns={0:"UID",1:"TAXID",2:"iseuk"}).set_index("UID")
uid2taxid_iseuk_df

kogs_domains = pd.merge(uid2taxid_iseuk_df, uid2kogs_df, on = "UID", how = "left")

kogs_domains["isprok"] = ~kogs_domains["iseuk"] 
has_prok = kogs_domains.reset_index().loc[:,["KOG","isprok"]].groupby("KOG").apply(lambda x: x.eq(True).any()).loc[:,["isprok"]].rename(columns={"isprok":"hasprok"})
has_euk = kogs_domains.reset_index().loc[:,["KOG","iseuk"]].groupby("KOG").apply(lambda x: x.eq(True).any()).loc[:,["iseuk"]].rename(columns={"iseuk":"haseuk"})


kogs_domains = kogs_domains.reset_index().set_index("KOG")

kogs_hasdomains = pd.merge(kogs_domains, has_prok, on="KOG", how = "left")

kogs_hasdomains = pd.merge(kogs_hasdomains, has_euk, on="KOG", how = "left")

kogs_hasboth = kogs_hasdomains.loc[:,["hasprok", "haseuk"]].reset_index().drop_duplicates()

kogs_hasboth["hasboth"] = kogs_hasboth["hasprok"] & kogs_hasboth["haseuk"]
kogs_hasboth["haseither"] = kogs_hasboth["hasprok"] | kogs_hasboth["haseuk"]

kogs_hasboth[kogs_hasboth["KOG"] == "K26400"]

KOGS_w_both = kogs_hasboth[kogs_hasboth["hasboth"] == True]["KOG"]
KOGS_w_both.to_csv("/data/luojaa/uid_stats/KOGs_w_both.csv")

sum(kogs_hasboth["hasprok"])

sum(kogs_hasboth["haseuk"])

sum(kogs_hasboth["hasboth"])

kogs_df = pd.merge(uid2kogs_df, genes2uid_df, on="UID", how = "outer")
kogs_df = pd.merge(kogs_df, uid2taxid_df, on = "UID", how = "outer")
kogs_df = pd.merge(kogs_df, uid2desc_df, on="UID", how = "outer")
# ENTRY is from KEGG gene scraping, and many of these genes don't have UIDs
# 50% genes don't have "KOGs" for this reason, so fill in gaps with "ENTRY"
# Also, ~600k "KOGs" map to UIDs that were "discovered", or not originally mapped from the KEGG gene scraping 
kogs_df["KOGID"] = kogs_df["KOG"].fillna(kogs_df["ENTRY"]) 

kogs_df["isnull"] = kogs_df.index.isnull()
null_count = kogs_df.reset_index().loc[:,["isnull", "KOGID"]].groupby("KOGID").sum().astype(int).reset_index().rename(columns={"isnull":"UIDS_MISSING"})
kogs_df = kogs_df.reset_index()
kogs_df = pd.merge(kogs_df, null_count, on = "KOGID", how = "outer")

kogs_df


kogs_df.to_csv(base_stats + "kogs_df.tsv", sep = "\t", index=None)

### map kogs to 3M uids that we include before realignment (eukaryotic)

euk_uid2kogs = "/data/luojaa/uid_stats/euk_uid2kogs.csv"

# kog_fastas_path = "/data/luojaa/kog_fastas_backup/"
# kog_fastas = os.listdir(kog_fastas_path)
# with open("/data/luojaa/uid_stats/uid2kogs_fasta.csv", "w") as outfile:
#     for file in kog_fastas:
#         kog = file.strip(".fasta")
#         with open(kog_fastas_path + file, "r") as handle:
#             for record in SeqIO.parse(handle, 'fasta'):
#                 uid = record.id
#                 print(",".join([uid, kog]), file=outfile)

euk_uid2kogs_df = pd.read_csv(euk_uid2kogs, header = None).rename(columns = {0:"UID", 1:"KOGID"}).set_index("UID")
euk_uid2kogs_df["ISEUK"] = [True] * len(euk_uid2kogs_df)
euk_uid2kogs_df = euk_uid2kogs_df.loc[:,"ISEUK"]

kogs_df_summary_in = pd.read_csv(base_stats + "kogs_df.tsv", sep = "\t").set_index("UID")

kogs_df_iseuk = pd.merge(kogs_df_summary_in, euk_uid2kogs_df, on = "UID", how="outer")
kogs_df_iseuk.fillna({"ISEUK":False}, inplace = True)
euk_count = kogs_df_iseuk.reset_index().loc[:,["ISEUK", "KOGID"]].groupby("KOGID").sum().astype(int).reset_index().rename(columns={"ISEUK":"EUKCOUNT"})

kogs_df_iseuk = kogs_df_iseuk.reset_index()
kogs_df_summary = pd.merge(kogs_df_iseuk, euk_count, on = "KOGID", how = "outer")

ncbi = ete3.NCBITaxa()

taxids = list(kogs_df_summary[~kogs_df_summary["TAXID"].isnull()]["TAXID"])
taxid2name = ncbi.get_taxid_translator(taxids)

def translate_taxid(name, dic):
    try:
        return dic[name]
    except:
        return np.nan

kogs_df_taxa = kogs_df_summary["TAXID"].apply(lambda x: translate_taxid(x, taxid2name))

kogs_df_crucial = kogs_df_summary.loc[:, ["UID", "KOGID", "FUNCTION", "ALIAS", "KEGG_ORG", "ISEUK"]]
kogs_df_crucial["TAXID"] = kogs_df_summary["TAXID"].fillna(0).astype(int)
kogs_df_crucial["SPECIES"] = kogs_df_taxa
kogs_df_searchable = kogs_df_crucial.fillna("none")

kogs_df_crucial.to_csv(base_stats + "kogs_df_crucial.tsv", sep = "\t", index=None)
kogs_df_searchable.to_csv(base_stats + "kogs_df_searchable.tsv", sep = "\t", index=None)

kogs_df_searchable

### format cluster_stats

fasta_stats_df = kogs_df_summary.loc[:,["KOGID", "UID", "UIDS_MISSING", "EUKCOUNT"]].groupby(["KOGID","UIDS_MISSING", "EUKCOUNT"]).count().reset_index().rename(columns={"UID":"UID_COUNT"}).set_index("KOGID")
cluster_size = pd.DataFrame(kogs_df_summary.loc[:,["KOGID"]].groupby("KOGID").size()).rename(columns={0:"KOGSIZE"})
fasta_stats_out = pd.merge(cluster_size, fasta_stats_df, on = "KOGID", how = "outer").reset_index()
fasta_stats_out["EUK_FRACTION"] = 100 * fasta_stats_out["EUKCOUNT"]/fasta_stats_out["UID_COUNT"]

fasta_stats_out[fasta_stats_out["EUKCOUNT"].between(1,10)]

fasta_stats_out.to_csv("/data/luojaa/uid_stats/cluster_stats.csv", sep = ",", index = None)

### format/explore  kegg "CATEGORIES"
to do: stack on pathway/module/reaction info

kegg_pathways = pd.read_csv("/data/luojaa/kegg/kegg_pathways.58345.tsv", sep = "\t").loc[:,["ENTRY", "PATHWAY_ID", "PATHWAY_NAME"]].rename(
    columns = {"PATHWAY_ID":"CATEGORY_ID", "PATHWAY_NAME":"CATEGORY_NAME", "ENTRY":"KOGID"})

kegg_rxns = pd.read_csv("/data/luojaa/kegg/kegg_reactions.tsv", sep = "\t").rename(columns = {"REACTION_ID":"CATEGORY_ID", "REACTION_NAME":"CATEGORY_NAME", "ENTRY":"KOGID"})
kegg_modules = pd.read_csv("/data/luojaa/kegg/kegg_modules.tsv", sep = "\t").rename(columns = {"MODULE_ID":"CATEGORY_ID", "MODULE_NAME":"CATEGORY_NAME", "ENTRY":"KOGID"})

# category: pathway/module/reaction; CID: map#####, CNAME: ___

kegg_categories_df = pd.concat([kegg_pathways, kegg_rxns, kegg_modules], axis = 0).dropna().reset_index().iloc[:, 1:]

kegg_categories_df.to_csv("/data/luojaa/kegg/kegg_categories.tsv", sep = "\t", index = None)

### make kogs for each category easily retrievable

kegg_categories = pd.read_csv("/data/luojaa/kegg/kegg_categories.tsv", sep = "\t")

kegg_categories["KOGIDS"] = kegg_categories.groupby(["CATEGORY_NAME", "CATEGORY_ID"])["KOGID"].transform(lambda x: ",".join(x))

kegg_categories["NUM_KOGS"] = kegg_categories.groupby(["CATEGORY_ID"]).transform(lambda x: len(x))["KOGID"]

kegg_categories_groups = kegg_categories.loc[:,["CATEGORY_NAME", "CATEGORY_ID", "KOGIDS", "NUM_KOGS"]].drop_duplicates().dropna()

kegg_categories

kegg_categories_groups.to_csv("/data/luojaa/kegg/kegg_categories_searchable.tsv", sep = "\t", index = None)

getkogs_wdescr(kogs_df_searchable, "malate")

# explore dataframes
def filter_iseuk_kogid(df, kogid):
    return df[(df["KOGID"] == kogid) & (df["ISEUK"])]
#filter_iseuk_kogid(kogs_df_searchable, "K00029")
def getkogs_wdescr(df, descr):
    return df[df["FUNCTION"].str.contains(descr)]["KOGID"].value_counts()
#getkogs_wdescr(kogs_df_searchable, "malate dehydrogenase")
def getkogs_wcategory(df, category):
    df_filtered = df[df["CATEGORY_NAME"].str.contains(category)]
    kogs_s = df_filtered["KOGIDS"].str.split(",")
    names_s = df_filtered["CATEGORY_NAME"]
    return dict(zip(names_s, kogs_s))
getkogs_wcategory(kegg_categories_groups, "malate")
