==> KEGG/09.rerun_swarm.ipynb <==
import os
import sys
module_path = os.path.abspath(os.path.join('..'))
if module_path not in sys.path:
    sys.path.append(module_path)

from core_functions.mmseqs_functions import realign_all_fastas_swarm_binned
from paths_and_parameters import realignment_swarm_opts

from core_functions.helper_functions import fasta_to_dict, dict_to_fasta, filter_by_entropy

# if the realignment script fails, K#####.fasta files missing "-"
# other realigned fasta files may also lack "-"
# these files may already by modified (cropped). This is good. Copy all such files from working directory to a new directory and rerun realignment script OR 

# if certain files fail to run or are super slow, just stop that swarm job and famsa the whole batch. 

# filter famsa alignments by entropy!

os.listdir("/data/luojaa/famsa_fastas")

base_root = "/data/luojaa/famsa_fastas/"

famsa_fastas = [f for f in os.listdir(base_root) if ".famsa" in f]

for famsa in famsa_fastas:
    filepath = base_root + famsa
    aln = fasta_to_dict(file=filepath)
    aln_filter = filter_by_entropy(aln, .25)
    outfasta = filepath + ".filtered"
    dict_to_fasta(aln_filter, write_file=outfasta)

# aln = fasta_to_dict(file=muscle_fasta)
# aln_filter = filter_by_entropy(aln, filter_entropy)

# if save_intermediate_files:
#     dict_to_fasta(aln_filter, write_file=f'{fasta}.aln')

# else:
#     dict_to_fasta(aln_filter, write_file=fasta)

#### scratch 

grep -H "-" /data/luojaa/kog_fastas/*.fasta | awk -F : ' { print $1 } ' | uniq | wc -l
13858

ls /data/luojaa/kog_fastas/*.fasta | wc -l
15658

there are cases where there are multiple maxxcc mergings per fasta, so this gives us an upper bound. the top command gives a lower bound.

some maxxcc mergings also don't result in completed runs (no "Wrote to fasta" following mergings). /data/luojaa/kog_fastas/K19326.fasta

for f in $(grep -L "-" /data/luojaa/kog_fastas/*.fasta); do cp $f /data/luojaa/unaligned_fastas/; done
