==> KEGG/11.summarize_search.ipynb <==
import pandas as pd

hhblits_results = "/data/luojaa/kegg_db/search/kegg_euk.tsv"
kog_descriptors = "/data/luojaa/uid_stats/kogs_df_crucial.tsv"


results = pd.read_csv(hhblits_results, sep = "\t")

descriptors = pd.read_csv(kog_descriptors, sep = "\t")

descriptors_df = descriptors.loc[:, ["KOGID", "FUNCTION"]].rename(columns={"FUNCTION": "header", "KOGID": "Target"}).dropna().set_index("Target")
descriptors_df = descriptors_df[~descriptors_df.index.duplicated(keep='first')]
descriptors_df

results_df = results.loc[:, ["Target", "Query", "Prob", "Pairwise_cov"]].rename(columns={"Query":"Target", "Target":"Query"}).set_index("Target")

results_annot = pd.merge(results_df, descriptors_df, on = "Target", how = "left").reset_index().loc[:, ["Query", "Target", "Prob", "Pairwise_cov", "header"]]

outfile = "/data/luojaa/kegg_db/search/kegg_euk.KEGG_annotation.tsv"
results_annot.to_csv(outfile, index = None, sep = "\t")

results_annot
